+++ Engine/inputManager.cpp
@@ -157,7 +157,7 @@ float InputManager::GetAxisAbsolute(int id, eDevice device, int index)
 		axis = m_mousePosition[id];
 		break;
 	case eDevice::CONTROLLER:
-		assert(index < m_controllers.size());
+		assert(index < (int)m_controllers.size());
 		axis = m_controllers[index].axis[id];
 		break;
 	}
@@ -178,7 +178,7 @@ float InputManager::GetAxisRelative(int id, eDevice device, int index)
 		axis = m_mousePosition[id] - m_prevMousePosition[id];
 		break;
 	case eDevice::CONTROLLER:
-		assert(index < m_controllers.size());
+		assert(index < (int)m_controllers.size());
 		axis = m_controllers[index].axis[id] - m_controllers[index].axis[id];
 		break;
 	}
@@ -199,7 +199,7 @@ bool InputManager::GetButtonDown(int id, eDevice device, int index)
 		buttonDown = m_mouseButtonstate & SDL_BUTTON(id);
 		break;
 	case eDevice::CONTROLLER:
-		assert(index < m_controllers.size());
+		assert(index < (int)m_controllers.size());
 		buttonDown = m_controllers[index].buttonstate[id];
 		break;
 	}
@@ -220,7 +220,7 @@ bool InputManager::GetPreviousButtonDown(int id, eDevice device, int index)
 		buttonDown = m_prevMouseButtonstate & SDL_BUTTON(id);
 		break;
 	case eDevice::CONTROLLER:
-		assert(index < m_controllers.size());
+		assert(index < (int)m_controllers.size());
 		buttonDown = m_controllers[index].prevButtonstate[id];
 		break;
 	}
